pipelines/model_training.py
"""
import whisper
import numpy as np
import librosa
import os
from scipy.spatial.distance import cosine

# Load Whisper model
model = whisper.load_model("base")

def transcribe_audio(file_path):
    Transcribe audio using Whisper.
    result = model.transcribe(file_path)
    return result["text"]

def extract_audio_features(file_path):
    Extract MFCC features from audio for comparison.
    y, sr = librosa.load(file_path, sr=16000)
    mfcc = librosa.feature.mfcc(y=y, sr=sr, n_mfcc=13)
    return np.mean(mfcc, axis=1)

def authenticate_voice(stored_voice, input_voice_path):
    Compare stored voice features with input voice for authentication.
    stored_features = np.frombuffer(stored_voice, dtype=np.float32)
    input_features = extract_audio_features(input_voice_path)
    
    similarity = 1 - cosine(stored_features, input_features)
    return similarity > 0.8  # Threshold for authentication

def analyze_sentiment(file_path):
    Perform sentiment analysis on transcribed text.
    transcription = transcribe_audio(file_path)
    
    # Placeholder sentiment analysis logic
    if any(word in transcription.lower() for word in ["happy", "great", "excited"]):
        sentiment = "Positive"
    elif any(word in transcription.lower() for word in ["sad", "upset", "angry"]):
        sentiment = "Negative"
    else:
        sentiment = "Neutral"
    
    return sentiment
"""